Add arterycode, signal number support in location search
by adding methods for these to `LocationSearchDAO`, plus a new `LocationSearchDAO.getSuggestions` method that stitches these together with textual intersection search.
