+++ lib/controller/LocationController.js
@@ -1,6 +1,7 @@
 import Joi from '@/lib/model/Joi';
 
 import CentrelineDAO from '@/lib/db/CentrelineDAO';
+import LocationSearchDAO from '@/lib/db/LocationSearchDAO';
 import { CentrelineType } from '@/lib/Constants';
 
 /**
@@ -38,8 +39,7 @@ LocationController.push({
   },
   handler: async (request) => {
     const { limit, q } = request.query;
-    // TODO: expand this to include signals and other features
-    return CentrelineDAO.intersectionSuggestions(q, limit);
+    return LocationSearchDAO.getSuggestions(q, limit);
   },
 });
